refactor: Clarify constant name
I think describing it as an estimate rather than an average makes more
sense because it's only an average based on the current set of stretches.
If we add or remove stretches, it's no longer an average; it's just an estimate.

I also think it's clearer to use "say" instead of "announce" because the
time is based on the time it took for the `say` utility to say it.