refactor: Replace quotes with more normal quotes
The double quotes I had originally were pasted from somewhere and
they weren't the same as the quotes I get when I hit the double quote
key in my editor. They did work, but they appear to be some other type
of double quote, so I thought it'd be best to replace them.